+++ ENGINE/src/utils.cpp
@@ -44,7 +44,7 @@ bool HasProjects()
 
 rl::Rectangle GetButtonSize(Types::Button& button)
 {
-	rl::Vector2 sizeinfo = rl::MeasureTextEx(button.font, button.text.c_str(), button.fontsize, 0.2f);
+	rl::Vector2 sizeinfo = rl::MeasureTextEx(button.font, button.text.c_str(), (int)button.fontsize, 0.2f);
 	rl::Rectangle rect = { button.position.x, button.position.y, sizeinfo.x * 2, sizeinfo.y * 2 };
 
 	float scaleX = (float)rl::GetScreenWidth() / (float)text.texture.width;
